Sock Merchant
John works at a clothing store. He has a large pile of socks that he must pair by color for sale. Given an array of integers representing the color of each sock, determine how many pairs of socks with matching colors there are.

For example, there are  N =7 socks with colors  ar = {1,2,1,2,1,2,3} There is one pair of color 1 and one of color  2 There are three odd socks left, one of each color. The number of pairs is  2